Extracting out tokenization alignment functions (#531)
* extracted ou talignment based tokenization funcs to src

* extracted ou talignment based tokenization funcs to src

* deleted unecessary import

* fixed tokenizer name

* refactored retokenized_edge_data.py
